refactor(nestjs-tools-file-storage): move types in separate files
